Refactor unique tests
