+++ py-polars/tests/unit/operations/unique/test_unique.py
@@ -1,0 +1,98 @@
+from datetime import date
+
+import polars as pl
+from polars.testing import assert_frame_equal, assert_series_equal
+
+
+def test_unique_predicate_pd() -> None:
+    lf = pl.LazyFrame(
+        {
+            "x": ["abc", "abc"],
+            "y": ["xxx", "xxx"],
+            "z": [True, False],
+        }
+    )
+
+    result = (
+        lf.unique(subset=["x", "y"], maintain_order=True, keep="last")
+        .filter(pl.col("z"))
+        .collect()
+    )
+    expected = pl.DataFrame(schema={"x": pl.String, "y": pl.String, "z": pl.Boolean})
+    assert_frame_equal(result, expected)
+
+    result = (
+        lf.unique(subset=["x", "y"], maintain_order=True, keep="any")
+        .filter(pl.col("z"))
+        .collect()
+    )
+    expected = pl.DataFrame({"x": ["abc"], "y": ["xxx"], "z": [True]})
+    assert_frame_equal(result, expected)
+
+
+def test_unique_on_list_df() -> None:
+    assert pl.DataFrame(
+        {"a": [1, 2, 3, 4, 4], "b": [[1, 1], [2], [3], [4, 4], [4, 4]]}
+    ).unique(maintain_order=True).to_dict(as_series=False) == {
+        "a": [1, 2, 3, 4],
+        "b": [[1, 1], [2], [3], [4, 4]],
+    }
+
+
+def test_list_unique() -> None:
+    s = pl.Series("a", [[1, 2], [3], [1, 2], [4, 5], [2], [2]])
+    assert s.unique(maintain_order=True).to_list() == [[1, 2], [3], [4, 5], [2]]
+    assert s.arg_unique().to_list() == [0, 1, 3, 4]
+    assert s.n_unique() == 4
+
+
+def test_unique_and_drop_stability() -> None:
+    # see: 2898
+    # the original cause was that we wrote:
+    # expr_a = a.unique()
+    # expr_a.filter(a.unique().is_not_null())
+    # meaning that the a.unique was executed twice, which is an unstable algorithm
+    df = pl.DataFrame({"a": [1, None, 1, None]})
+    assert df.select(pl.col("a").unique().drop_nulls()).to_series()[0] == 1
+
+
+def test_unique_empty() -> None:
+    for dt in [pl.String, pl.Boolean, pl.Int32, pl.UInt32]:
+        s = pl.Series([], dtype=dt)
+        assert_series_equal(s.unique(), s)
+
+
+def test_unique() -> None:
+    ldf = pl.LazyFrame({"a": [1, 2, 2], "b": [3, 3, 3]})
+
+    expected = pl.DataFrame({"a": [1, 2], "b": [3, 3]})
+    assert_frame_equal(ldf.unique(maintain_order=True).collect(), expected)
+
+    result = ldf.unique(subset="b", maintain_order=True).collect()
+    expected = pl.DataFrame({"a": [1], "b": [3]})
+    assert_frame_equal(result, expected)
+
+    s0 = pl.Series("a", [1, 2, None, 2])
+    # test if the null is included
+    assert s0.unique().to_list() == [None, 1, 2]
+
+
+def test_struct_unique_df() -> None:
+    df = pl.DataFrame(
+        {
+            "numerical": [1, 2, 1],
+            "struct": [{"x": 1, "y": 2}, {"x": 3, "y": 4}, {"x": 1, "y": 2}],
+        }
+    )
+
+    df.select("numerical", "struct").unique().sort("numerical")
+
+
+def test_sorted_unique_dates() -> None:
+    assert (
+        pl.DataFrame(
+            [pl.Series("dt", [date(2015, 6, 24), date(2015, 6, 23)], dtype=pl.Date)]
+        )
+        .sort("dt")
+        .unique()
+    ).to_dict(as_series=False) == {"dt": [date(2015, 6, 23), date(2015, 6, 24)]}
